Move all legacy home stuff into legacy folder
