+++ src/plugins/home/public/application/components/legacy/add_data/add_data.tsx
@@ -30,58 +30,53 @@
 
 import React, { FC } from 'react';
 import PropTypes from 'prop-types';
-import { EuiFlexGroup, EuiHorizontalRule, EuiSpacer, EuiTitle, EuiFlexItem } from '@elastic/eui';
+import { EuiFlexGroup, EuiFlexItem, EuiSpacer, EuiTitle } from '@elastic/eui';
 import { FormattedMessage } from '@osd/i18n/react';
 // @ts-expect-error untyped service
-import { FeatureCatalogueEntry } from '../../services';
-import { createAppNavigationHandler } from '../app_navigation_handler';
+import { FeatureCatalogueEntry } from '../../../services';
+import { createAppNavigationHandler } from '../../app_navigation_handler';
+
 // @ts-expect-error untyped component
-import { Synopsis } from '../synopsis';
+import { Synopsis } from '../../synopsis';
 
 interface Props {
   addBasePath: (path: string) => string;
   features: FeatureCatalogueEntry[];
 }
 
-export const ManageData: FC<Props> = ({ addBasePath, features }) => (
-  <>
-    {features.length > 1 && <EuiHorizontalRule margin="xl" aria-hidden="true" />}
-
-    {features.length > 0 && (
-      <section
-        className="homDataManage"
-        aria-labelledby="homDataManage__title"
-        data-test-subj="homDataManage"
-      >
+export const AddData: FC<Props> = ({ addBasePath, features }) => (
+  <section className="homDataAdd" aria-labelledby="homDataAdd__title">
+    <EuiFlexGroup alignItems="center">
+      <EuiFlexItem grow={1}>
         <EuiTitle size="s">
-          <h2 id="homDataManage__title">
-            <FormattedMessage id="home.manageData.sectionTitle" defaultMessage="Manage your data" />
+          <h2 id="homDataAdd__title">
+            <FormattedMessage id="home.addData.sectionTitle" defaultMessage="Ingest your data" />
           </h2>
         </EuiTitle>
+      </EuiFlexItem>
+    </EuiFlexGroup>
 
-        <EuiSpacer size="m" />
+    <EuiSpacer size="m" />
 
-        <EuiFlexGroup className="homDataManage__content">
-          {features.map((feature) => (
-            <EuiFlexItem key={feature.id}>
-              <Synopsis
-                id={feature.id}
-                onClick={createAppNavigationHandler(feature.path)}
-                description={feature.description}
-                iconType={feature.icon}
-                title={feature.title}
-                url={addBasePath(feature.path)}
-                wrapInPanel
-              />
-            </EuiFlexItem>
-          ))}
-        </EuiFlexGroup>
-      </section>
-    )}
-  </>
+    <EuiFlexGroup className="homDataAdd__content">
+      {features.map((feature) => (
+        <EuiFlexItem key={feature.id}>
+          <Synopsis
+            id={feature.id}
+            onClick={createAppNavigationHandler(feature.path)}
+            description={feature.description}
+            iconType={feature.icon}
+            title={feature.title}
+            url={addBasePath(feature.path)}
+            wrapInPanel
+          />
+        </EuiFlexItem>
+      ))}
+    </EuiFlexGroup>
+  </section>
 );
 
-ManageData.propTypes = {
+AddData.propTypes = {
   addBasePath: PropTypes.func.isRequired,
   features: PropTypes.arrayOf(
     PropTypes.shape({
