+++ src/plugins/home/public/application/components/legacy/manage_data/index.tsx
