+++ src/plugins/home/public/application/components/legacy/manage_data/manage_data.test.tsx
@@ -29,10 +29,10 @@
  */
 
 import React from 'react';
-import { AddData } from './add_data';
+import { ManageData } from './manage_data';
 import { shallowWithIntl } from 'test_utils/enzyme_helpers';
 
-jest.mock('../app_navigation_handler', () => {
+jest.mock('../../app_navigation_handler', () => {
   return {
     createAppNavigationHandler: jest.fn(() => () => {}),
   };
@@ -46,42 +46,58 @@ const addBasePathMock = jest.fn((path: string) => (path ? path : 'path'));
 
 const mockFeatures = [
   {
-    category: 'data',
-    description: 'Get started with sample data, visualizations, and dashboards.',
+    category: 'admin',
+    description: 'Control who has access and what tasks they can perform.',
+    icon: 'securityApp',
+    id: 'security',
+    order: 600,
+    path: 'path-to-security-roles',
+    title: 'Protect your data',
     showOnHomePage: true,
-    icon: 'indexOpen',
-    id: 'home_tutorial_directory',
-    order: 500,
-    path: '/app/home#/tutorial_directory',
-    title: 'Ingest data',
   },
   {
     category: 'admin',
-    description: 'Add and manage your fleet of OpenSearch Agents and integrations.',
+    description: 'Track the real-time health and performance of your deployment.',
+    icon: 'monitoringApp',
+    id: 'monitoring',
+    order: 610,
+    path: 'path-to-monitoring',
+    title: 'Monitor the stack',
     showOnHomePage: true,
-    icon: 'indexManagementApp',
-    id: 'ingestManager',
-    order: 510,
-    path: '/app/ingestManager',
-    title: 'Add OpenSearch Agent',
   },
   {
-    category: 'data',
-    description: 'Import your own CSV, NDJSON, or log file',
+    category: 'admin',
+    description:
+      'Save snapshots to a backup repository, and restore to recover index and cluster state.',
+    icon: 'storage',
+    id: 'snapshot_restore',
+    order: 630,
+    path: 'path-to-snapshot-restore',
+    title: 'Store & recover backups',
     showOnHomePage: true,
-    icon: 'document',
-    id: 'ml_file_data_visualizer',
-    order: 520,
-    path: '/app/ml#/filedatavisualizer',
-    title: 'Upload a file',
+  },
+  {
+    category: 'admin',
+    description: 'Define lifecycle policies to automatically perform operations as an index ages.',
+    icon: 'indexSettings',
+    id: 'index_lifecycle_management',
+    order: 640,
+    path: 'path-to-index-lifecycle-management',
+    title: 'Manage index lifecycles',
+    showOnHomePage: true,
   },
 ];
 
-describe('AddData', () => {
+describe('ManageData', () => {
   test('render', () => {
     const component = shallowWithIntl(
-      <AddData addBasePath={addBasePathMock} features={mockFeatures} />
+      <ManageData addBasePath={addBasePathMock} features={mockFeatures} />
     );
+    expect(component).toMatchSnapshot();
+  });
+
+  test('render empty without any features', () => {
+    const component = shallowWithIntl(<ManageData addBasePath={addBasePathMock} features={[]} />);
     expect(component).toMatchSnapshot();
   });
 });
