+++ src/plugins/home/public/application/components/legacy/solutions_section/index.ts
