+++ src/plugins/home/public/application/components/legacy/solutions_section/solution_panel.test.tsx
