+++ src/plugins/home/public/application/components/legacy/welcome.test.tsx
@@ -31,10 +31,10 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import { Welcome } from './welcome';
-import { telemetryPluginMock } from '../../../../telemetry/public/mocks';
-import { getLogosMock } from '../../../../../core/common/mocks';
+import { telemetryPluginMock } from '../../../../../telemetry/public/mocks';
+import { getLogosMock } from '../../../../../../core/common/mocks';
 
-jest.mock('../opensearch_dashboards_services', () => ({
+jest.mock('../../opensearch_dashboards_services', () => ({
   getServices: () => ({
     addBasePath: (path: string) => `root${path}`,
     trackUiMetric: () => {},
